refactor: 충돌 및 스타일 수정
