+++ src/components/ui/TextSize.tsx
@@ -2,7 +2,7 @@ import { Text, TextStyle } from "react-native";
 
 interface TextSizeProps {
   size: number;
-  color: string;
+  color?: string;
   content: string;
   weight?: TextStyle["fontWeight"];
 }
